:: removed timer.cpp and timer.h, the timer is now integrated into the main/oscillator files. The timer is working as it should be for now, and changes the frequency according to the array that is created by the melody generator. ::

+++ CSD2b/Synthsong/melodygen.h
@@ -8,7 +8,7 @@ class Melodygen
 public:
   Melodygen();
   ~Melodygen();
-  float* generate(int amount);
+  double* generate(int amount);
   float mtoF(int midiNote);
 private:
   int amount;
